Rename files and add in lofi credit
