+++ src/App.tsx
@@ -3,14 +3,14 @@ import { ThemeProvider } from '@mui/material'
 import { theme } from './utils/customMUITheme'
 import PomodoroTimer from './components/PomodoroTimer'
 import TodoList from './components/TodoList'
-import LofiPlayer from './components/LofiPlayer'
+import StickyButtons from './components/StickyButtons'
 
 export default function App() {
   return (
     <ThemeProvider theme={theme}>
       <PomodoroTimer />
       <TodoList />
-      <LofiPlayer />
+      <StickyButtons />
     </ThemeProvider>
   )
 }
