--- src/components/LofiPlayer.less
@@ -1,5 +1,0 @@
-.lofi-player {
-  position: fixed;
-  bottom: 16px;
-  right: 16px;
-}
