+++ src/components/StickyButtons.less
@@ -1,0 +1,17 @@
+.sticky-buttons {
+  position: fixed;
+  bottom: 16px;
+  right: 16px;
+}
+
+.lofi-player-information {
+  position: fixed;
+  bottom: 16px;
+  left: 16px;
+
+  .paragraph {
+    font-size: 14px;
+    margin: 0;
+    text-align: left;
+  }
+}
